% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_3d_reductions.R
\name{add_all_projections}
\alias{add_all_projections}
\alias{add_2d_tsne}
\alias{add_2d_umap}
\alias{add_3d_tsne}
\alias{add_3d_umap}
\title{Calculate tSNE and UMAP projections in 2- and 3-dimensions}
\usage{
add_all_projections(seurat, n_dimensions = 40, input_reduction = "pca")

add_2d_tsne(seurat, n_dimensions = 40, input_reduction = "pca")

add_2d_umap(seurat, n_dimensions = 40, input_reduction = "pca")

add_3d_tsne(seurat, n_dimensions = 40, input_reduction = "pca")

add_3d_umap(seurat, n_dimensions = 40, input_reduction = "pca")
}
\arguments{
\item{seurat}{Seurat object}

\item{n_dimensions}{Number of dimensions from which the projection is calculated}

\item{input_reduction}{The reduction from which the projection is calculated}
}
\description{
Calculate tSNE and UMAP projections in 2- and 3-dimensions

Calculate tSNE in 2-dimensions

Calculate UMAP in 2-dimensions

Calculate tSNE in 3-dimensions

Calculate UMAP in 3-dimensions
}
\section{Functions}{
\itemize{
\item \code{add_all_projections}: Add 3D tSNE reduction to Seurat object

\item \code{add_2d_tsne}: Add 2D tSNE reduction to Seurat object

\item \code{add_2d_umap}: Add 3D UMAP reduction to Seurat object

\item \code{add_3d_tsne}: Add 3D tSNE reduction to Seurat object

\item \code{add_3d_umap}: Add 3D UMAP reduction to Seurat object
}}

