% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_cluster_identity_sets.R
\name{write_cluster_identity_sets}
\alias{write_cluster_identity_sets}
\alias{guess_cluster_identity_sets}
\title{Write cell cluster sets to a h5 file}
\usage{
write_cluster_identity_sets(h5_file, cluster_identity_sets, ...)

guess_cluster_identity_sets(seurat)
}
\arguments{
\item{h5_file}{Path to \code{h5} file}

\item{cluster_identity_sets}{A \code{data.frame} of meta data}

\item{...}{Arguments passed to \code{guess_cluster_identity_sets}}

\item{seurat}{Seurat object}
}
\description{
Write cell cluster sets to a h5 file

Get cluster identity sets
}
\details{
Selects \verb{_snn_res} variables as cluster identity sets and selects all identities by default.
}
\section{Functions}{
\itemize{
\item \code{write_cluster_identity_sets}: Write cluster identities list to \code{h5} file

\item \code{guess_cluster_identity_sets}: Select Seurat cluster identities
}}

