% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_to_scv.R
\name{create_h5_scv}
\alias{create_h5_scv}
\title{Create an empty \code{h5} file}
\usage{
create_h5_scv(h5_file, delete = TRUE)
}
\arguments{
\item{h5_file}{Path to \code{h5} file}

\item{remove}{Should \code{h5_file} be removed if it exists?}
}
\description{
Create an empty \code{h5} file
}
