% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{get_config_values}
\alias{get_config_values}
\title{Parse configuration object to get parameter values}
\usage{
get_config_values(config, key)
}
\arguments{
\item{config}{The parsed configuration file}

\item{key}{The key to extract}
}
\description{
Parse configuration object to get parameter values
}
\details{
Searches in \code{config} to find the most-dataset-specific value for \code{key}.
}
