% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{get_dataset_choices}
\alias{get_dataset_choices}
\title{Parse configuration object to get available datasets}
\usage{
get_dataset_choices(config, datasets = config$datasets)
}
\arguments{
\item{config}{The parsed configuration file}

\item{datasets}{The nested list of datasets, a subset of \code{config}}
}
\description{
Parse configuration object to get available datasets
}
