% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{load_app_config}
\alias{load_app_config}
\title{Read the configuration yaml}
\usage{
load_app_config(file = "config.yaml")
}
\arguments{
\item{file}{Path to YAML}
}
\description{
Read the configuration yaml
}
