% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_selectors.R
\name{preferred_choice}
\alias{preferred_choice}
\title{Make a ranked choice}
\usage{
preferred_choice(x, preferences, default = 1)
}
\arguments{
\item{x}{Named vector of possible choices}

\item{preferences}{Ranked vector of names of \code{x}}

\item{default}{Element of \code{x} if none of \code{preferences} are names of \code{x}}
}
\description{
Make a ranked choice
}
