% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scviewer.R
\name{run_scviewer}
\alias{run_scviewer}
\title{Start scviewer}
\usage{
run_scviewer(...)
}
\arguments{
\item{...}{All arguments are ignored}
}
\description{
Start scviewer
}
\details{
Uses the \code{ui} and \code{server} functions defined in \code{scviewer}.
}
