% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{scViewer}
\alias{scViewer}
\title{Run scViewer shiny app}
\usage{
scViewer(dev = FALSE)
}
\arguments{
\item{dev}{Run the applicaiton in developer mode}
}
\value{
The shiny app will open
}
\description{
Run scViewer shiny app
}
\examples{
\dontrun{
scViewer()
}
}
