% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_to_scv.R
\name{seurat_to_scv}
\alias{seurat_to_scv}
\title{Convert a Seurat object to a scviewer-formatted \code{h5} file}
\usage{
seurat_to_scv(h5_file, seurat, recalculate_reductions = FALSE, ...)
}
\arguments{
\item{h5_file}{Path to \code{h5} file}

\item{seurat}{Seurat object or file path from which an \code{rds} can be read}

\item{recalculate_reductions}{Should 2D and 3D tSNE and UMAP reductions be calculated?}

\item{...}{Arguments passed to \code{add_all_projections}}
}
\description{
Convert a Seurat object to a scviewer-formatted \code{h5} file
}
