% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_message.R
\name{skip_log_message}
\alias{skip_log_message}
\alias{show_log_message}
\title{Do not print log messages}
\usage{
skip_log_message(...)

show_log_message(text, prepend = "///")
}
\arguments{
\item{...}{All arguments are ignored}

\item{text}{Message to output}

\item{prepend}{Maker to print in front of message}
}
\description{
Do not print log messages

Print log messages
}
\section{Functions}{
\itemize{
\item \code{skip_log_message}: Hide log messages

\item \code{show_log_message}: Print log messages
}}

\concept{log messages}
