% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_cell_filters.R
\name{write_cell_filter_parameters}
\alias{write_cell_filter_parameters}
\title{Write cell filter parameters to a h5 file}
\usage{
write_cell_filter_parameters(h5_file, cell_filter_parameters)
}
\arguments{
\item{h5_file}{Path to \code{h5} file}

\item{cell_filter_parameters}{A \code{list} of filter parameters (see details)}
}
\description{
Write cell filter parameters to a h5 file
}
\details{
The cell filter parameters list is named by the UI element label and contains:
\itemize{
\item \code{var} the variable from \code{metadata} (which should be a factor) that is to be filtered
\item \code{selected} the levels that are selected by default
}
}
