% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_features.R
\name{write_features}
\alias{write_features}
\alias{guess_features_matrix}
\title{Write feature values to a h5 file}
\usage{
write_features(h5_file, features_matrix, ...)

guess_features_matrix(seurat)
}
\arguments{
\item{h5_file}{Path to \code{h5} file}

\item{features_matrix}{Matrix of cells (rows) and features (columns)}

\item{...}{Arguments passed to \code{guess_features_matrix}}

\item{seurat}{Seurat object}
}
\description{
Write feature values to a h5 file

Get a feature values matrix
}
\details{
Combines the \code{RNA@data} slot and any numeric meta data variables into a \code{feature_matrix}.
}
\section{Functions}{
\itemize{
\item \code{write_features}: Write feature matrix to \code{h5_file}

\item \code{guess_features_matrix}: Create a feature values matrix using default features
}}

