% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_metadata.R
\name{write_metadata}
\alias{write_metadata}
\alias{guess_metadata}
\title{Write metadata to a h5 file}
\usage{
write_metadata(h5_file, metadata, ...)

guess_metadata(seurat)
}
\arguments{
\item{h5_file}{Path to \code{h5} file}

\item{metadata}{A \code{data.frame} of meta data}

\item{...}{Arguments passed to \code{guess_metadata}}

\item{seurat}{Seurat object}
}
\description{
Write metadata to a h5 file

Get dataset meta data
}
\details{
Extracts the factor- or character-type variables from the \code{seurat} meta data, but it uses \code{!is.numeric}!
}
\section{Functions}{
\itemize{
\item \code{write_metadata}: Write meta data to \code{h5} file

\item \code{guess_metadata}: Select non-numeric variables from the meta data slot
}}

