% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{scViewer}
\alias{scViewer}
\title{Run scViewer shiny app}
\usage{
scViewer(filelocation = NULL)
}
\arguments{
\item{filelocation}{The directory of sce object or seurat object. If not specific the file can be uploaded inside the app.}
}
\value{
The shiny app will open
}
\description{
Run scViewer shiny app
}
