% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_features.R
\name{write_features}
\alias{write_features}
\alias{guess_features_matrix}
\title{Write feature values to a h5 file}
\usage{
write_features(h5_file, feature_types, features_matrix, ..., dry_run = FALSE)

guess_features_matrix(seurat, ...)
}
\arguments{
\item{h5_file}{Path to \code{h5} file}

\item{feature_types}{A list of feature names indexed by feature type}

\item{features_matrix}{Matrix of cells (rows) and features (columns)}

\item{...}{Ignored}

\item{seurat}{Seurat object}
}
\description{
Write feature values to a h5 file

Get a feature values matrix
}
\details{
The \code{feature_names} list uses the feature type to influence types of plot and app behaviour so should be one of those recognised by the app. Defaults to \code{any} if not specified.

Combines the \code{RNA@data} slot and any numeric meta data variables into a \code{feature_matrix}.
}
\section{Functions}{
\itemize{
\item \code{write_features}: Write feature matrix to \code{h5_file}

\item \code{guess_features_matrix}: Create a feature values matrix using default features
}}

